chore(c-protocol): move C implementation to subdirectory
