+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ repos:
       - id: no-commit-to-branch
         stages: [pre-commit, manual]
   - repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: v20.1.0
+    rev: v20.1.4
     hooks:
       - id: clang-format
         stages: [pre-commit, manual]
