+++ C/eai_protocol_init.cmake
