+++ C/src/communicationEndpoint/CMakeLists.txt
