+++ C/src/communicationEndpoint/include/public/eaip/endpoint/CommunicationEndpoint.h
